Factor out svgs
