+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/new-flow/ContributionThankYouCompleteRegistration.jsx
@@ -11,6 +11,7 @@ import ActionHeader from './components/ActionHeader';
 import ActionBody from './components/ActionBody';
 import ExpandableContainer from './components/ExpandableContainer';
 import BulletPointedList from './components/BulletPointedList';
+import SvgPersonWithTick from './components/SvgPersonWithTick';
 import styles from './styles';
 
 const bodyText = css`
@@ -32,23 +33,6 @@ const form = css`
 const buttonContainer = css`
   margin-top: ${space[6]}px;
 `;
-
-const SvgPersonWithTick = () => (
-  <svg
-    width="39"
-    height="38"
-    viewBox="0 0 39 38"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M17.4733 16.4666C19.785 16.4666 22.5083 13.8699 22.5083 10.7033C22.5083 7.53661 20.64 5.69995 17.4733 5.69995C14.3067 5.69995 12.47 7.53661 12.47 10.7033C12.47 13.8699 15.415 16.4666 17.4733 16.4666ZM24.7566 18.9999L26.1183 17.6699L27.5116 19.0949L31.6599 14.9466L32.9899 16.3082L27.5116 21.7866L26.1816 20.4249L24.7566 18.9999ZM17.4733 18.9999C18.835 18.9999 20.1333 19.0949 21.305 19.3166C21.6533 22.6099 24.25 25.2699 27.48 25.8399L28.8733 31.0332L27.575 32.2999H7.30836L6.07336 31.0332L8.57502 21.5332L9.87335 20.2349C12.4067 19.3799 14.6867 18.9999 17.4733 18.9999Z"
-      fill="#121212"
-    />
-  </svg>
-);
 
 const ContributionThankYouContinueToAccount = () => {
   const [isExpanded, setIsExpanded] = useState(false);
