+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/new-flow/ContributionThankYouSetSupportReminder.jsx
@@ -12,6 +12,7 @@ import ActionContainer from './components/ActionContainer';
 import ActionHeader from './components/ActionHeader';
 import ActionBody from './components/ActionBody';
 import ExpandableContainer from './components/ExpandableContainer';
+import SvgClock from './components/SvgClock';
 import styles from './styles';
 
 const bodyText = css`
@@ -47,23 +48,6 @@ const privacyTextLink = css`
 
   color: ${neutral[20]};
 `;
-
-const SvgClock = () => (
-  <svg
-    width="39"
-    height="38"
-    viewBox="0 0 39 38"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M33.4332 18.9999C33.4332 26.6949 27.1949 32.9332 19.4999 32.9332C11.805 32.9332 5.56665 26.6949 5.56665 18.9999C5.56665 11.305 11.805 5.06665 19.4999 5.06665C27.1949 5.06665 33.4332 11.305 33.4332 18.9999ZM21.0199 18.5566L20.1016 7.59998H18.8666L17.9166 19.4749L19.5316 21.0899L28.3666 20.2666V18.9999L21.0199 18.5566Z"
-      fill="#121212"
-    />
-  </svg>
-);
 
 const ContributionThankYouSetSupportReminder = () => {
   const [isExpanded, setIsExpanded] = useState(false);
