+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/new-flow/components/SvgLinkedIn.jsx
@@ -1,0 +1,15 @@
+// @flow
+import React from 'react';
+
+const SvgLinkedIn = () => (
+  <svg
+    width="31"
+    height="31"
+    viewBox="0 0 31 31"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M8.78303 5.5C9.37125 5.52578 9.92684 5.77799 10.334 6.20406C10.7411 6.63013 10.9684 7.19717 10.9684 7.78697C10.9684 8.37677 10.7411 8.94381 10.334 9.36988C9.92684 9.79595 9.37125 10.0482 8.78303 10.0739C8.17753 10.0739 7.59683 9.83299 7.16868 9.4041C6.74053 8.97521 6.5 8.39351 6.5 7.78697C6.5 7.18043 6.74053 6.59873 7.16868 6.16984C7.59683 5.74095 8.17753 5.5 8.78303 5.5ZM6.81367 11.809H10.7524V24.5H6.81367V11.809ZM13.2225 11.809H16.9996V13.5441H17.0535C17.5789 12.5454 18.8642 11.4926 20.7797 11.4926C24.7635 11.4926 25.5 14.1215 25.5 17.5392V24.5H21.5671V18.3295C21.5671 16.8568 21.538 14.9643 19.5206 14.9643C17.4712 14.9643 17.1583 16.5681 17.1583 18.223V24.5H13.2225V11.809Z" />
+  </svg>
+);
+
+export default SvgLinkedIn;
