+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/new-flow/components/SvgShare.jsx
@@ -1,0 +1,21 @@
+// @flow
+import React from 'react';
+
+const SvgShare = () => (
+  <svg
+    width="39"
+    height="38"
+    viewBox="0 0 39 38"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M27.0999 24.0666C29.2216 24.0666 30.8999 25.7766 30.8999 27.8666C30.8999 29.9882 29.2216 31.6666 27.0999 31.6666C25.0099 31.6666 23.3316 29.9882 23.3316 27.8666C23.3316 27.7082 23.3316 27.5182 23.3633 27.3916L12.0266 21.7233C11.2983 22.4199 10.4116 22.7683 9.3983 22.7683C7.24498 22.7683 5.56665 21.0899 5.56665 18.9999C5.56665 16.8783 7.24498 15.1683 9.3983 15.1683C10.38 15.1683 11.2983 15.6116 12.0266 16.2766L23.3633 10.6083C23.3316 10.4816 23.3316 10.3233 23.3316 10.1016C23.3316 8.01164 25.0099 6.33331 27.0999 6.33331C29.2216 6.33331 30.8999 8.01164 30.8999 10.1016C30.8999 12.2233 29.2216 13.9333 27.0999 13.9333C26.0866 13.9333 25.2316 13.5216 24.5033 12.8883L13.135 18.5249C13.1666 18.6199 13.1666 18.7783 13.1666 18.9999C13.1666 19.2216 13.1666 19.3799 13.135 19.4749L24.5033 25.1116C25.2316 24.4783 26.0866 24.0666 27.0999 24.0666ZM27.0999 7.85331C25.8649 7.85331 24.8199 8.83497 24.8199 10.1016C24.8199 11.3683 25.8649 12.4133 27.0999 12.4133C28.3666 12.4133 29.4116 11.3683 29.4116 10.1016C29.4116 8.83497 28.3666 7.85331 27.0999 7.85331ZM27.0999 30.1466C28.3666 30.1466 29.4116 29.1016 29.4116 27.8666C29.4116 26.5999 28.3666 25.5866 27.0999 25.5866C25.8649 25.5866 24.8199 26.5999 24.8199 27.8666C24.8199 29.1016 25.8649 30.1466 27.0999 30.1466Z"
+      fill="#121212"
+    />
+  </svg>
+);
+
+export default SvgShare;
